# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface Design {
  id: number;
  name: string;
  /** Price in dollars */
  price: number;
  /** Duration of the style in hours */
  durationHours: number;
  /** URL of the design image */
  imageUrl?: string | null;
  description?: string | null;
  createdAt: string;
}

export interface CreateDesignBody {
  name: string;
  price: number;
  durationHours: number;
  imageUrl?: string | null;
  description?: string | null;
}

export interface AvailabilityRange {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time in HH:MM format (24h) */
  startTime: string;
  /** End time in HH:MM format (24h) */
  endTime: string;
  createdAt: string;
}

export interface CreateAvailabilityBody {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time in HH:MM (24h) */
  startTime: string;
  /** End time in HH:MM (24h) */
  endTime: string;
}

export interface TimeSlot {
  /** Start time in HH:MM (24h) */
  startTime: string;
  /** End time in HH:MM (24h) */
  endTime: string;
}

export interface AvailableSlotsResponse {
  date: string;
  designId: number;
  designName: string;
  durationHours: number;
  slots: TimeSlot[];
}

export interface Booking {
  id: number;
  customerName: string;
  customerPhone?: string | null;
  customerEmail?: string | null;
  designId: number;
  designName: string;
  /** Booking date YYYY-MM-DD */
  date: string;
  /** Booking start HH:MM (24h) */
  startTime: string;
  /** Booking end HH:MM (24h) */
  endTime: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateBookingBody {
  customerName: string;
  customerPhone?: string | null;
  customerEmail?: string | null;
  designId: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time in HH:MM (24h) */
  startTime: string;
  notes?: string | null;
}

export type AdminSummaryPopularDesignsItem = {
  designId: number;
  designName: string;
  bookingCount: number;
};

export interface AdminSummary {
  totalBookings: number;
  totalDesigns: number;
  upcomingBookings: number;
  todayBookings: number;
  popularDesigns: AdminSummaryPopularDesignsItem[];
}

export type ListAvailabilityParams = {
  /**
   * Filter by specific date (YYYY-MM-DD)
   */
  date?: string;
};

export type GetAvailableSlotsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
  /**
   * ID of the selected design
   */
  designId: number;
};

export type ListBookingsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  designId?: number;
};

export type GetUpcomingBookingsParams = {
  limit?: number;
};