# lib/api-zod/src/generated/types/availabilityRange.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilityRange {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time in HH:MM format (24h) */
  startTime: string;
  /** End time in HH:MM format (24h) */
  endTime: string;
  createdAt: Date;
}