# lib/api-zod/src/generated/types/availableSlotsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { TimeSlot } from "./timeSlot";

export interface AvailableSlotsResponse {
  date: string;
  designId: number;
  designName: string;
  durationHours: number;
  slots: TimeSlot[];
}