# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  customerName: string;
  customerPhone?: string | null;
  customerEmail?: string | null;
  designId: number;
  designName: string;
  /** Booking date YYYY-MM-DD */
  date: string;
  /** Booking start HH:MM (24h) */
  startTime: string;
  /** Booking end HH:MM (24h) */
  endTime: string;
  notes?: string | null;
  createdAt: Date;
}