# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  customerName: string;
  customerPhone?: string | null;
  customerEmail?: string | null;
  designId: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Start time in HH:MM (24h) */
  startTime: string;
  notes?: string | null;
}