# lib/api-zod/src/generated/types/createDesignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDesignBody {
  name: string;
  price: number;
  durationHours: number;
  imageUrl?: string | null;
  description?: string | null;
}