# lib/api-zod/src/generated/types/design.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Braids Center Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface Design {
  id: number;
  name: string;
  /** Price in dollars */
  price: number;
  /** Duration of the style in hours */
  durationHours: number;
  /** URL of the design image */
  imageUrl?: string | null;
  description?: string | null;
  createdAt: Date;
}